+++ lib/Component/LegacyException.php
@@ -1,12 +1,11 @@
 <?php
-/**
- * Exception raised by Legacy
- */
-
 namespace Lmc\Steward\Component;
 
 use Exception;
 
+/**
+ * Exception raised by Legacy
+ */
 class LegacyException extends \Exception
 {
     public function __construct($message = "", $code = 0, Exception $previous = null)